chore: remove RCTAddressSanitizerCrashManager (#2395)
## Summary:

This module is flagged internally by our static analysis. This module
and test was added to test that ASAN was actually enabled and would
crash on unsafe code. This meant checking in unsafe code as a test,
which is now flagged now that we build from source internally. We don't
need this module and test anymore, and we never had any automated tests
that actually checked that this crashes. Let's just delete it.

## Test Plan:

CI should pass.