fix(macos): <Text> with backgroundColor spills color past it's frame (#2401)
## Summary:

For some odd reason, if you set `backgroundColor` on a Text component,
the color spills out past the frame to the parent. We can work around
this by setting border radius... which somehow prevents this behavior. I
need to debug the native side more (and may update this PR with a proper
fix if I find one), but for now let's patch the JS to set a dummy value
for borderRadius if we create a Text Element with backgroundColor and
without borderRadius

## Test Plan:

A bunch of examples in RNTester are now visible... like the "Recently
Viewed" section.

<img width="616" alt="image"
src="https://github.com/user-attachments/assets/10ac2e6e-2ee6-4a04-8f96-56c92dcc2e68"
/>