chore(macOS): Consolidate RCTAccessibilityManager between iOS and macOS (#2400)
## Summary:

This PR attempts to solve two problems:

1. We have had a forked AccessibilityManager module for macOS for a
while. In general, I prefer ifdefing rather than forking so that we get
changes to the iOS code during merges. Let's combine the files.

2. I want to cleanup macOS's "highContrast" API in AccessibilityInfo so
it can be upstreamed, so that RNW can follow it and we have a
cross-platform way to get high contrast state. Luckily, the API we have
for macOS can be ported to iOS, so I did that. Unluckily, someone
already did that (see:
https://github.com/facebook/react-native/pull/46826 ) and used the iOS
specific name "Darker system colors" instead of something like "increase
contrast" or "high contrast". So I'll probably need to add my API
upstream, deprecate that one, and then remove it in a future release.